refactor: auth/member controller and service separation
