+++ src/main/java/com/spacestar/back/member/controller/MemberController.java
@@ -34,31 +34,7 @@ public class MemberController {
 
     private final MemberService memberService;
     private final ModelMapper mapper;
-    private final JWTUtil jwtUtil;
 
-    @Operation(summary = "회원가입")
-    @PostMapping("/join")
-    public ResponseEntity<Void> memberInfoAdd(@RequestBody @Valid MemberJoinReqVo memberJoinReqVo) {
-
-        memberService.addMember(mapper.map(memberJoinReqVo, MemberJoinReqDto.class));
-        return new ResponseEntity<>(ResponseSuccess.SIGNUP_SUCCESS);
-    }
-
-    @Operation(summary = "닉네임 중복 검증")
-    @GetMapping("/duplication/{nickname}")
-    public ResponseEntity<NicknameResVo> duplicationNickname(@PathVariable("nickname") String nickname) {
-        return new ResponseEntity<>(ResponseSuccess.DUPLICATION_NICKNAME_SUCCESS,mapper.map(memberService.duplicationNickname(nickname), NicknameResVo.class));
-    }
-
-    @Operation(summary = "카카오 로그인")
-    @PostMapping("/login")
-    public ResponseEntity<HttpHeaders> login(@RequestBody MemberLoginReqVo memberLoginReqVo){
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.set(HttpHeaders.AUTHORIZATION, memberService.kakaoLogin(mapper.map(memberLoginReqVo, MemberLoginReqDto.class)).getAccessToken());
-
-        return new ResponseEntity<>(ResponseSuccess.LOGIN_SUCCESS, headers);
-    }
 
     @Operation(summary = "회원 정보 추가입력/ 회원 정보 수정")
     @PutMapping("/info/add")
@@ -81,4 +57,5 @@ public class MemberController {
 
         return new ResponseEntity<>(ResponseSuccess.PROFILE_IMAGE_UPDATE_SUCCESS);
     }
+
 }
